feat: implement CLI manager and system orchestrator with comprehensive logging and error handling
