+++ pyproject.toml
@@ -92,10 +92,6 @@ skip = [
   "tests/fake_project",
 ]
 
-[tool.flake8-annotations]
-suppress-none-returning = true
-suppress-dummy-args = true
-allow-untyped-defs = true
 
 [tool.bandit]
 skips = ["B101"]
